Move definition of TreeConstruction::calculateInternalNodesBoundingVolumes() to declaration file
